pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aurarouter"
version = "0.2.0"
authors = [{ name = "Steven Siebert", email = "info@auracoredynamics.com" }]
description = "Multi-model MCP routing fabric for local and cloud LLMs"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
dependencies = [
    "mcp[cli]",
    "httpx",
    "google-genai",
    "pyyaml",
    "anthropic",
    "llama-cpp-python>=0.3.0",
    "huggingface-hub>=0.20.0",
]

[project.optional-dependencies]
gui = ["PySide6>=6.6"]
dev = ["pytest", "pytest-mock", "pytest-cov", "ruff"]
all = ["PySide6>=6.6", "pytest", "pytest-mock", "pytest-cov", "ruff"]

[project.scripts]
aurarouter = "aurarouter.cli:main"
aurarouter-gui = "aurarouter.gui.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
